Delegate trace commands to plugin runtime
